refactor(bridge): collapse TaskStore base class into TaskRunStore
Eliminate single-subclass inheritance hierarchy by folding TaskStore directly into TaskRunStore. Production only ever instantiated TaskRunStore, and TaskRunStore did not override any base methods.

- Move persistence, CRUD, and context reconciliation logic into TaskRunStore
- Update test/task-store.test.js to instantiate TaskRunStore directly
- Remove bridge/src/task-store.js
